Rename logos for consistency
